+++ src/pe_scanner/integration/__init__.py
@@ -39,3 +39,4 @@ __all__ = [
     "sync_with_master",
 ]
 
+